Le combat vraiment pas fini

+++ DoorMonsterLoot/Assets/Scripts/Battle/projectile.cs
@@ -6,7 +6,7 @@ public class projectile : MonoBehaviour {
 
     public bool gogogo = false;
     public int sens = 0;
-    public float dommageSiLeJoueurEstTouche = 0;
+    public int dommageSiLeJoueurEstTouche = 0;
 
 	// Use this for initialization
 	void Start () {
@@ -25,20 +25,17 @@ public class projectile : MonoBehaviour {
             else
             {
                 gameObject.transform.position = new Vector2(gameObject.transform.position.x, gameObject.transform.position.y - Time.deltaTime * 20);
-
             }
         }
-
 
         if (gameObject.transform.position.x <= -15 || gameObject.transform.position.y <= -10)
         {
             Debug.Log("Je suis deleter" + Time.time);
             Destroy(gameObject);
         }
-
 	}
 
-    public float retournerValeurDeDommage()
+    public int retournerValeurDeDommage()
     {
         return dommageSiLeJoueurEstTouche;
     }
